perf: Tweaks to reduce cycle consumptions
This commit introduces 2 changes based on real observations to reduce
cycle consumptions:

* FastAlloc would process the whole buffer into a double-linked list
at initialization phase, however, a smart contract might only use a
handful of the Nodes there. This commit changes the code so we are
only initializating a constant number of nodes, and only create more
nodes into the double-linked list when we needed. This also has the
benefit that a bigger FastAlloc buffer can be used without cycle
penalty.
* BuddyAlloc would always do memset on the memory buffer, however, in
most cases, the buffer fed to BuddyAlloc is already zero-filled(such
as ckb-std), this commit as a new param so we can skip the memset in
BuddyAlloc, saving us the cycles used to memset buffers, which are not
actually needed.